[`flake8-future-annotations`] Add autofix (`FA100`)
Summary
--

This PR resolves the easiest part of #18502 by adding an autofix that just adds
`from __future__ import annotations` at the top of the file. We couldn't use any
of the usual, existing `Importer` methods because the import must be at the very
beginning of the file, so I factored out `Importer::add_import_at_start_of_file`
from `Importer::add_import`.

I think this is technically an unsafe fix because it changes a lot of behavior,
but that's also the point of the rule, so I made it a safe fix for now. I'm
happy to change it to unsafe, though.

I also expected the fix only to be available sometimes, but `add_import` and
`add_import_at_start_of_file` are actually infallible.

Test Plan
--

Existing snapshots, updated to add the fixes.